what even is life?

+++ TerrainGeneration/src/TerrainGenerator.cpp
@@ -6,7 +6,7 @@
 #include "TerrainGenerator.h"
 
 
-TerrainGenerator::TerrainGenerator(size_t _dimension) : m_dimension(_dimension) {}
+TerrainGenerator::TerrainGenerator(int _dimension) : m_dimension(_dimension) {}
 
 
 void TerrainGenerator::generate()
@@ -20,20 +20,20 @@ void TerrainGenerator::generate()
 
   m_heightMap = {};
 
-  for (size_t i =0; i<m_dimension*m_dimension; ++i)
+  for (int i =0; i<m_dimension*m_dimension; ++i)
   {
-    float height = perlinModule.GetValue(getX(i)/10,getY(i)/10,m_seed);
+    float height = float(perlinModule.GetValue(getX(i)/10,getY(i)/10,m_seed));
     m_heightMap.push_back(m_amplitude*height);
   }
 }
 
 
-float TerrainGenerator::getX(const int _index) const
+double TerrainGenerator::getX(const int _index) const
 {
   return _index%m_dimension;
 }
 
-float TerrainGenerator::getY(const int _index) const
+double TerrainGenerator::getY(const int _index) const
 {
   return _index/m_dimension;
 }
